requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
python-dotenv==1.0.1
langchain==0.3.14
langchain-community==0.3.13
langchain-core==0.3.29
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.4